+++ CHANGELOG.md
@@ -14,6 +14,7 @@ accidentally triggering the load of a previous DB version.**
 * #1007 Remove parent oid from find_children_oids result
 * #1038 Infer time_bucket_gapfill arguments from WHERE clause
 * #1067 Add treat_null_as_missing option to locf
+* #1134 Add support for partitionwise aggregation
 
 **Bugfixes**
 * #1115 Fix ordered append optimization for join queries